pyproject.toml
[project]
name = "modguard-rl"
version = "1.0.0"
description = "Synthetic content-policy triage RL environment"
requires-python = ">=3.11"

dependencies = [
  "openenv-core",
  "fastapi",
  "uvicorn[standard]",
  "pydantic>=2.0",
  "httpx",
  "openai",
  "python-dotenv",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "client"]
